+++ docs/changelog.rst
@@ -18,7 +18,8 @@ CHANGELOG
 - Fix: 'length' filter can now be displayed on mobile apps
 - Fix: make 'duration' mobile filter consistent in doc
 - Fix: Label annotation becoming empty when double-clicked (refs #3927)
-
+- Fix: missing default values in SQL template for Report model (#4085)
+- Fix: users cannot add attachment to Report model (#4085)
 
 2.107.0 (2024-06-07)
 ------------------------