+++ docs/changelog.rst
@@ -8,6 +8,8 @@ CHANGELOG
 **Improvements**
 
 - Allows user to edit flatpage WYSIWYG button-link and suggestion block
+- Display years in Report filter in descending order (#4085)
+- Change CSS overflow in filters popup, to display full select dropdowns (with mapentity update)
 
 **Documentation**
 
@@ -18,6 +20,9 @@ CHANGELOG
 - Fix: 'length' filter can now be displayed on mobile apps
 - Fix: make 'duration' mobile filter consistent in doc
 - Fix: Label annotation becoming empty when double-clicked (refs #3927)
+- Fix: missing default values in SQL template for Report model (#4085)
+- Fix: users cannot add attachment to Report model (#4085)
+- Fix: remove legacy default value for `external_uuid` in database for Report model (#4085)
 
 
 2.107.0 (2024-06-07)