+++ docs/changelog.rst
@@ -18,6 +18,7 @@ CHANGELOG
 - Fix: 'length' filter can now be displayed on mobile apps
 - Fix: make 'duration' mobile filter consistent in doc
 - Fix: Label annotation becoming empty when double-clicked (refs #3927)
+- Fix: Too long linestring can't store elevation chart in default cache
 
 
 2.107.0 (2024-06-07)