+++ docs/changelog.rst
@@ -14,6 +14,7 @@ CHANGELOG
 - Fix empty linetring in reorder_topology cmd (fixes #4092)
 - Prevent Pillow Decompressed Data Too Large Error from interrupting Parsers
 - Fix ZeroDivisionError raised on generating thumbnails in APIv2 (#4231)
+- Fix error on APIv2 for Signages with pictograms (#3839)
 
 
 2.108.0     (2024-07-12)